+++ Spring Boot/src/main/java/wi/pb/clothesshop/service/OrderService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface OrderService {
 
-    Order placeOrder(Long userId);
+    Order placeOrder(int userId);
     Order getOrder(Long orderId);
 
-    List<Order> getUserOrders(Long userId);
+    List<Order> getUserOrders(int userId);
 }
